remove workflow_functions.py and move write_output_video function to utils.py
